requirements-dev.txt
pytest==8.0.0
pytest-cov==4.1.0
pytest-flask==1.3.0
black==24.1.1
isort==5.13.2
autoflake==2.2.1
pylint==3.0.3
mypy==1.8.0
types-requests>=2.31.0
